Finalized reading from database.
The entirety of the XML can be read and stored in memory. For the time being, however, the only thing that can be done without it is to output to the screen. To facilitate reading from the file, some changes were made to the file format. No longer will data values within a row in table have <v> tags; instead, each value will be stored on a single line. (This means that databases will not be able to hld the '\n' character. This may be revisited later.