Cleaned up current program.
Cleaned up the output of the current program to only output the table information.  Added 'free's for all of the arrays in the database struct and all of the 'strings.' Also added more data to the test_db.xml file to assure that the current program works with multiple tables.