Handled invalid input.
Handled the case for when a user inputs a non-existant table. When doing so, he is informed that the table does not exist and the next statement can be executed. Also added comments to make the reading of the WHERE clause clearer (because I already managed to confuse myself). Moved the printing code from an early version (which has remained commented out since then) to its own function so that it can be used for testing purposes, if needed.