Can extract fields from a SELECT statement.
Added the ability to extract the fields to SELECT in queries. Will likely have to be modified later (when there is a second set of statements with commas in them - the WHERE clause, that is - likely with pointer comparisons and such). But for now, the extraction works fine. (Right now they're just printed to the screen.)