Made SELECT * and SELECT field1, field2, ..., fieldN work
Added the ability to SELECT only specific fields. After extracting the fields with get_fields(), strcmp was used to create an array of the corresponding indices for these fields in the Table struct's fields array. A slight modification to create_return_string() made it possible to use it for both styles of SELECT statements. I also believe I fixed all of the memory leaking in slect_statement for now (though I will look back at it again once the program is finished).